Apply Implementaion pattern for better naming for interfaces
